refactor(macos): separate window and ambient background responsibilities
