+++ Sources/Kumone/Features/Shared/MainWindowAmbientBackground.swift
@@ -1,6 +1,8 @@
 import SwiftUI
 
 #if os(macOS)
+import AppKit
+
 /// A non-interactive artwork tint layered above the split view's opaque system
 /// surfaces. The low opacity keeps system materials and controls readable.
 struct MainWindowAmbientBackground: View {
@@ -59,6 +61,178 @@ enum MainWindowAmbientOpacity {
 
     static func glow(isDark: Bool, intensity: Double) -> Double {
         (isDark ? 0.18 : 0.12) * intensity
+    }
+}
+
+struct MainWindowAmbientConfiguration {
+    let showsAmbientBackground: Bool
+    let showsTitlebarAmbientBackground: Bool
+    let colors: ArtworkColors
+    let mainColumnWidth: CGFloat
+    let intensity: Double
+}
+
+/// Owns the AppKit state required to extend the artwork tint through the main
+/// window titlebar while preserving the window's original appearance.
+@MainActor
+final class MainWindowAmbientAppearanceController {
+    private var configuration = MainWindowAmbientConfiguration(
+        showsAmbientBackground: false,
+        showsTitlebarAmbientBackground: false,
+        colors: .fallback,
+        mainColumnWidth: 0,
+        intensity: 1
+    )
+    private var titlebarWasTransparent: Bool?
+    private var hadFullSizeContentView = false
+    private var titlebarMask: TitlebarMaskView?
+
+    func configure(_ configuration: MainWindowAmbientConfiguration, in window: NSWindow) {
+        let wasShowingAmbientBackground = self.configuration.showsAmbientBackground
+        self.configuration = configuration
+
+        guard wasShowingAmbientBackground != configuration.showsAmbientBackground else {
+            updateLayout(in: window)
+            return
+        }
+
+        if configuration.showsAmbientBackground {
+            titlebarWasTransparent = window.titlebarAppearsTransparent
+            hadFullSizeContentView = window.styleMask.contains(.fullSizeContentView)
+            updateLayout(in: window)
+        } else {
+            window.titlebarAppearsTransparent = titlebarWasTransparent ?? false
+            if hadFullSizeContentView {
+                window.styleMask.insert(.fullSizeContentView)
+            } else {
+                window.styleMask.remove(.fullSizeContentView)
+            }
+            titlebarWasTransparent = nil
+            titlebarMask?.removeFromSuperview()
+            titlebarMask = nil
+        }
+    }
+
+    func updateLayout(in window: NSWindow) {
+        guard configuration.showsAmbientBackground else { return }
+        if !window.titlebarAppearsTransparent {
+            window.titlebarAppearsTransparent = true
+        }
+        if !window.styleMask.contains(.fullSizeContentView) {
+            window.styleMask.insert(.fullSizeContentView)
+        }
+        guard configuration.showsTitlebarAmbientBackground else {
+            titlebarMask?.removeFromSuperview()
+            titlebarMask = nil
+            return
+        }
+        installTitlebarMask(in: window)
+        layoutTitlebarMask(in: window)
+    }
+
+    private func installTitlebarMask(in window: NSWindow) {
+        guard titlebarMask == nil, let contentView = window.contentView else { return }
+        let mask = TitlebarMaskView()
+        contentView.addSubview(mask, positioned: .above, relativeTo: nil)
+        titlebarMask = mask
+    }
+
+    private func layoutTitlebarMask(in window: NSWindow) {
+        guard let mask = titlebarMask, let contentView = window.contentView else { return }
+        let width = min(max(configuration.mainColumnWidth, 0), contentView.bounds.width)
+        guard width > 0 else {
+            mask.isHidden = true
+            return
+        }
+
+        let titlebarHeight = contentView.bounds.height - window.contentLayoutRect.height
+        guard titlebarHeight > 0 else {
+            mask.isHidden = true
+            return
+        }
+        let y = contentView.isFlipped
+            ? contentView.bounds.minY
+            : contentView.bounds.maxY - titlebarHeight
+        let frame = CGRect(
+            x: contentView.bounds.maxX - width,
+            y: y,
+            width: width,
+            height: titlebarHeight
+        )
+
+        if mask.frame != frame {
+            mask.frame = frame
+        }
+        if mask.isHidden {
+            mask.isHidden = false
+        }
+        mask.update(
+            colors: configuration.colors,
+            appearance: window.effectiveAppearance,
+            intensity: configuration.intensity
+        )
+    }
+}
+
+private final class TitlebarMaskView: NSView {
+    private let gradientLayer = CAGradientLayer()
+    private var appliedColors: ArtworkColors?
+    private var appliedAppearanceName: NSAppearance.Name?
+    private var appliedIntensity: Double?
+
+    override init(frame frameRect: NSRect) {
+        super.init(frame: frameRect)
+        wantsLayer = true
+        layer?.addSublayer(gradientLayer)
+    }
+
+    @available(*, unavailable)
+    required init?(coder: NSCoder) { fatalError("init(coder:) is not used") }
+
+    override func layout() {
+        super.layout()
+        gradientLayer.frame = bounds
+    }
+
+    override func hitTest(_ point: NSPoint) -> NSView? { nil }
+
+    func update(colors: ArtworkColors, appearance: NSAppearance, intensity: Double) {
+        guard appliedColors != colors
+                || appliedAppearanceName != appearance.name
+                || appliedIntensity != intensity else {
+            return
+        }
+        appliedColors = colors
+        appliedAppearanceName = appearance.name
+        appliedIntensity = intensity
+
+        let isDark = appearance.bestMatch(from: [.aqua, .darkAqua]) == .darkAqua
+        let opacity = CGFloat(
+            MainWindowAmbientOpacity.gradient(isDark: isDark, intensity: intensity)
+        )
+        let base = NSColor.windowBackgroundColor
+        let primary = blended(NSColor(colors.primary), over: base, opacity: opacity)
+        let secondary = blended(NSColor(colors.secondary), over: base, opacity: opacity)
+
+        CATransaction.begin()
+        CATransaction.setDisableActions(Platform.isReduceMotionEnabled)
+        CATransaction.setAnimationDuration(0.6)
+        CATransaction.setAnimationTimingFunction(CAMediaTimingFunction(name: .easeInEaseOut))
+        gradientLayer.colors = [primary.cgColor, secondary.cgColor]
+        gradientLayer.startPoint = CGPoint(x: 0, y: 1)
+        gradientLayer.endPoint = CGPoint(x: 1, y: 0)
+        CATransaction.commit()
+    }
+
+    private func blended(_ color: NSColor, over base: NSColor, opacity: CGFloat) -> NSColor {
+        let foreground = color.usingColorSpace(.extendedSRGB) ?? color
+        let background = base.usingColorSpace(.extendedSRGB) ?? base
+        return NSColor(
+            red: background.redComponent + (foreground.redComponent - background.redComponent) * opacity,
+            green: background.greenComponent + (foreground.greenComponent - background.greenComponent) * opacity,
+            blue: background.blueComponent + (foreground.blueComponent - background.blueComponent) * opacity,
+            alpha: 1
+        )
     }
 }
 #endif
